MsBuild: get rid of unneeded rebuilds
Fix MsBuild's change detection to get rid of unneeded project rebuilds.

MsBuild can match inputs and outputs one-to-one, but for this to work
transform must be put directly into `Outputs` parameter of the target.

This change caused `Compile` items in the target to be added twice so
they were moved to a separate target which works fine.